Update sitl-native-on-windows.rst
Goal is to provide the reader the option to use Jsbsim 

The instructions for the added section were copied from the unix version of this page (https://github.com/ArduPilot/ardupilot_wiki/edit/master/dev/source/docs/setting-up-sitl-on-linux.rst). 

I verified that the unix instructions work perfectly fine on Cygwin, which means a jsbsim based model (i.e. Rascal:jsbsim or a custom one) SITL simulation can be run on a windows machine.